ml_service/requirements.txt
torch
torchvision
torchcam
flask
scikit-learn
Pillow
wandb